resolve more circular import issues because of moving frame attributes (merge conflict)
