+++ astropy/coordinates/builtin_frames/ecliptic.py
@@ -4,7 +4,8 @@ from __future__ import (absolute_import, unicode_literals, division,
                         print_function)
 
 from ..representation import SphericalRepresentation
-from ..baseframe import BaseCoordinateFrame, TimeFrameAttribute
+from ..baseframe import BaseCoordinateFrame
+from ..frame_attributes import TimeFrameAttribute
 from .utils import EQUINOX_J2000
