+++ astropy/coordinates/builtin_frames/gcrs.py
@@ -5,9 +5,9 @@ from __future__ import (absolute_import, unicode_literals, division,
 
 from ... import units as u
 from ..representation import SphericalRepresentation
-from ..baseframe import (BaseCoordinateFrame, RepresentationMapping,
-                         TimeFrameAttribute,
-                         CartesianRepresentationFrameAttribute)
+from ..baseframe import BaseCoordinateFrame, RepresentationMapping
+from ..frame_attributes import (TimeFrameAttribute,
+                                CartesianRepresentationFrameAttribute)
 from .utils import DEFAULT_OBSTIME, EQUINOX_J2000
