+++ astropy/coordinates/builtin_frames/hcrs.py
@@ -4,8 +4,8 @@ from __future__ import (absolute_import, unicode_literals, division,
                         print_function)
 
 from ..representation import SphericalRepresentation
-from ..baseframe import (BaseCoordinateFrame, RepresentationMapping,
-                         TimeFrameAttribute)
+from ..baseframe import BaseCoordinateFrame, RepresentationMapping
+from ..frame_attributes import TimeFrameAttribute
 from .utils import DEFAULT_OBSTIME
