+++ astropy/coordinates/builtin_frames/skyoffset.py
@@ -6,9 +6,9 @@ from __future__ import (absolute_import, division, print_function)
 
 from ... import units as u
 from ..transformations import DynamicMatrixTransform, FunctionTransform
-from ..baseframe import (CoordinateAttribute, QuantityFrameAttribute,
-                         frame_transform_graph, RepresentationMapping,
+from ..baseframe import (frame_transform_graph, RepresentationMapping,
                          BaseCoordinateFrame)
+from ..frame_attributes import CoordinateAttribute, QuantityFrameAttribute
 from ..matrix_utilities import (rotation_matrix,
                                 matrix_product, matrix_transpose)
 from ...utils.compat import namedtuple_asdict
